Move state related tests to new module
