+++ tests/unit/entities/test_link.py
@@ -5,7 +5,6 @@ from typing import ContextManager, Iterable, Mapping
 
 import pytest
 
-from dj_link.entities import command
 from dj_link.entities.link import Transfer, create_link, pull
 from dj_link.entities.state import (
     Activated,
@@ -273,91 +272,3 @@ class TestPull:
         }
         actual = pull(link, requested={Identifier("1")})
         assert actual == expected
-
-
-def test_pulling_idle_entity_returns_correct_commands() -> None:
-    link = create_link(create_assignments({Components.SOURCE: {"1"}}))
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.pull() == {command.AddToOutbound(Identifier("1")), command.StartPullOperation(Identifier("1"))}
-
-
-def test_processing_activated_entity_undergoing_pull_operation_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}}),
-        operations={Operations.PULL: {Identifier("1")}},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.process() == {command.AddToLocal(Identifier("1"))}
-
-
-def test_processing_activated_entity_undergoing_delete_operation_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}}),
-        operations={Operations.DELETE: {Identifier("1")}},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.process() == {
-        command.RemoveFromOutbound(Identifier("1")),
-        command.FinishDeleteOperation(Identifier("1")),
-    }
-
-
-def test_processing_received_entity_undergoing_pull_operation_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}, Components.LOCAL: {"1"}}),
-        operations={Operations.PULL: {Identifier("1")}},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.process() == {command.FinishPullOperation(Identifier("1"))}
-
-
-def test_processing_received_entity_undergoing_delete_operation_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}, Components.LOCAL: {"1"}}),
-        operations={Operations.DELETE: {Identifier("1")}},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.process() == {command.RemoveFromLocal(Identifier("1"))}
-
-
-def test_deleting_pulled_entity_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}, Components.LOCAL: {"1"}})
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.delete() == {command.StartDeleteOperation(Identifier("1"))}
-
-
-def test_flagging_pulled_entity_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}, Components.LOCAL: {"1"}})
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.flag() == {command.Flag(Identifier("1"))}
-
-
-def test_unflagging_tainted_entity_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}, Components.LOCAL: {"1"}}),
-        tainted_identifiers={Identifier("1")},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.unflag() == {command.Unflag(Identifier("1"))}
-
-
-def test_deleting_tainted_entity_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}, Components.OUTBOUND: {"1"}, Components.LOCAL: {"1"}}),
-        tainted_identifiers={Identifier("1")},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.delete() == {command.StartDeleteOperation(Identifier("1"))}
-
-
-def test_unflagging_deprecated_entity_returns_correct_commands() -> None:
-    link = create_link(
-        create_assignments({Components.SOURCE: {"1"}}),
-        tainted_identifiers={Identifier("1")},
-    )
-    entity = next(iter(link[Components.SOURCE]))
-    assert entity.unflag() == {command.Unflag(Identifier("1"))}
